+++ src/rest.ts
@@ -2,3 +2,4 @@ export {ResponseType, LocationHeaderAndJsonResponse, LocationHeaderResponse, Blo
 export {Cache} from './rest/cache';
 export {Body,BaseUrl,DefaultHeaders,DELETE,GET,HEAD,Header,Headers,JsonContentType,ParameterTransform,Path,POST,Produces,PUT,Query,QueryObject,ResponseTransform,RESTClient,FullHttpResponse,ReportProgress, DisableInterceptor} from './rest/common';
 export {RestTransferStateService} from './transferState/restTransferState.service';
+export {AdditionalInfoPropertyDescriptor} from './rest/additionalInfoPropertyDescriptor';